+++ react-components/src/components/Form/TextArea/index.jsx
@@ -1,4 +1,4 @@
-import React, { memo, useRef } from 'react'
+import React, { memo } from 'react'
 import PropTypes from 'prop-types'
 
 import { FormGroup } from '../FormGroup'
@@ -25,7 +25,7 @@ export const TextArea = memo(
       if (el) {
         el.style.overflow = 'hidden'
         el.style.height = 0 // This is so the scroll-height doen't get padded out by a larget min-height
-        el.style.height = el.scrollHeight + 'px'
+        el.style.height = `${el.scrollHeight}px`
       }
     }
 
@@ -65,7 +65,10 @@ TextArea.propTypes = {
   value: PropTypes.string,
   onChange: PropTypes.func.isRequired,
   description: PropTypes.string,
-  tooltip: PropTypes.objectOf(PropTypes.string),
+  tooltip: PropTypes.shape({
+    content: PropTypes.string,
+    title: PropTypes.string,
+  }),
   example: PropTypes.oneOfType([
     PropTypes.shape({
       buttonTitle: PropTypes.string,
@@ -92,7 +95,7 @@ TextArea.defaultProps = {
   placeholder: '',
   value: '',
   description: '',
-  tooltip: {},
+  tooltip: null,
   example: {},
   hideLabel: false,
   lesson: {},