+++ react-components/tests/components/MarkLessonAsComplete.test.jsx
@@ -1,25 +1,26 @@
 import React from 'react'
-import Enzyme from 'enzyme'
-import Adapter from 'enzyme-adapter-react-16'
-import { act } from 'react-dom/test-utils'
+import { render, waitFor } from '@testing-library/react'
 import fetchMock from 'fetch-mock'
+
 import { MarkLessonAsComplete } from '@src/components/MarkLessonAsComplete/MarkLessonAsComplete'
-Enzyme.configure({ adapter: new Adapter() })
 
 const mockResponse = {
-  lesson_completed: [{}, {}]
+  lesson_completed: [{}, {}],
 }
 
 describe('MarkLessonAsComplete', () => {
-  let wrapper
-  const nonTickedText = 'Lesson complete?Yes'
-  const tickedText = 'Lesson complete?Great! Progress saved'
-
   beforeEach(() => {
-    fetchMock.get('http://localhost/sso/api/v1/lesson-completed/20/', mockResponse, { overwriteRoutes: false })
-    fetchMock.post('/sso/api/v1/lesson-completed/20/', mockResponse, { overwriteRoutes: false })
-    fetchMock.delete('/sso/api/v1/lesson-completed/20/', mockResponse, { overwriteRoutes: false })
-    wrapper = Enzyme.mount(<MarkLessonAsComplete endpoint="/sso/api/v1/lesson-completed/20/" />)
+    fetchMock.get(
+      'http://localhost/sso/api/v1/lesson-completed/20/',
+      mockResponse,
+      { overwriteRoutes: false }
+    )
+    fetchMock.post('/sso/api/v1/lesson-completed/20/', mockResponse, {
+      overwriteRoutes: false,
+    })
+    fetchMock.delete('/sso/api/v1/lesson-completed/20/', mockResponse, {
+      overwriteRoutes: false,
+    })
   })
 
   afterEach(() => {
@@ -27,33 +28,32 @@ describe('MarkLessonAsComplete', () => {
   })
 
   it('renders correct elements', async () => {
-    await act(async () => {
-      expect(wrapper.find('.mark-lesson-as-complete').length).toEqual(1)
-      expect(wrapper.find('h2').length).toEqual(1)
-      expect(wrapper.find('.great-checkbox').length).toEqual(1)
+    const { container } = render(
+      <MarkLessonAsComplete endpoint="/sso/api/v1/lesson-completed/20/" />
+    )
+
+    await waitFor(() => {
+      expect(
+        container.querySelectorAll('.mark-lesson-as-complete')
+      ).toHaveLength(1)
+      expect(container.querySelectorAll('h2')).toHaveLength(1)
+      expect(container.querySelectorAll('.great-checkbox')).toHaveLength(1)
     })
   })
 
-  it('updates label text', async () => {
-    await act(async () => {
-      expect(wrapper.find('label').text()).toEqual(nonTickedText)
-      await act(async () => {
-        wrapper
-          .find('input')
-          .props()
-          .onChange()
-      })
+  it('updates label text on click', async () => {
+    const { container } = render(
+      <MarkLessonAsComplete endpoint="/sso/api/v1/lesson-completed/20/" />
+    )
 
-      expect(wrapper.find('label').text()).toEqual(nonTickedText)
+    expect(container.querySelector('label').textContent).toMatch('Yes')
 
-      await act(async () => {
-        wrapper
-          .find('input')
-          .props()
-          .onClick()
-      })
+    container.querySelector('input').click()
 
-      expect(wrapper.find('label').text()).toEqual(tickedText)
+    await waitFor(() => {
+      expect(container.querySelector('label').textContent).toMatch(
+        'Great! Progress saved'
+      )
     })
   })
 })