+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - GLS-97 - fix and make header styles more consistent
 - GLS-124 - new homepage banner
+- NOTICKET - retire Enzyme in tests
 
 ## [2.6.0](https://github.com/uktrade/great-cms/releases/tag/2.6.0)